+++ source/renderer/app/containers/wallet/Wallet.js
@@ -65,11 +65,6 @@ export default class Wallet extends Component<Props> {
     const isRestoreActive =
       get(wallets, ['active', 'syncState', 'status']) ===
       WalletSyncStateStatuses.RESTORING;
-    const restoreProgress = get(
-      activeWallet,
-      ['syncState', 'progress', 'quantity'],
-      0
-    );
 
     const {
       recoveryPhraseVerificationStatus,
@@ -80,9 +75,7 @@ export default class Wallet extends Component<Props> {
 
     return (
       <MainLayout>
-        {isRestoreActive ? (
-          <RestoreNotification restoreProgress={restoreProgress} />
-        ) : null}
+        {isRestoreActive ? <RestoreNotification /> : null}
 
         <WalletWithNavigation
           isActiveScreen={this.isActiveScreen}